add_benchmarks

+++ src/optim/base.py
@@ -16,7 +16,7 @@ from .utils import eval, get_batch, save_checkpoint
 def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, batch_size, sequence_length, eval_freq, ckpt_path, distributed_backend,extra_args, itr=0,rng_state_dict=None):
     device_type = 'cuda' if 'cuda' in str(extra_args.device) else 'cpu'
     type_ctx = nullcontext() if device_type == 'cpu' else torch.amp.autocast(
-        device_type=device_type, dtype=torch.bfloat16)  # extra_args.dtype)
+        device_type=device_type, dtype=torch.float16)  # extra_args.dtype)
     best_val_loss, text_table = float('inf'), None # best_val_loss not used atm, early stopping not recommended but possible 
     substep = itr * acc_steps
     data["train"], train_sampler = get_dataloader(
